refactor: re-org Liquidator helpers
