+++ src/Liquidator/helpers/actionTypes.ts
@@ -1,0 +1,13 @@
+export enum ActionType {
+  OpenVault,
+  MintShortOption,
+  BurnShortOption,
+  DepositLongOption,
+  WithdrawLongOption,
+  DepositCollateral,
+  WithdrawCollateral,
+  SettleVault,
+  Redeem,
+  Call,
+  Liquidate,
+}
