+++ src/Liquidator/helpers/liquidation.ts
@@ -1,40 +1,17 @@
 import { ethers } from "ethers";
-import fetch from "node-fetch";
 
-import erc20ABI from "./erc20ABI";
+import { ActionType } from "./actionTypes";
+import { fetchDeribitETHIndexPrice } from "./deribit";
 import marginCalculatorABI from "./marginCalculatorABI";
-import oTokenABI from "./oTokenABI";
 import {
-  liquidatorAccount,
   gammaControllerProxyContract,
+  liquidatorAccount,
   provider,
-} from "../helpers";
-
-enum ActionType {
-  OpenVault,
-  MintShortOption,
-  BurnShortOption,
-  DepositLongOption,
-  WithdrawLongOption,
-  DepositCollateral,
-  WithdrawCollateral,
-  SettleVault,
-  Redeem,
-  Call,
-  Liquidate,
-}
+} from "../../helpers";
 
 const liquidatorAccountAddress = liquidatorAccount.address;
 
 const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
-
-const generateDeribitUrl = ({
-  expiryDate,
-  optionType,
-  strikePrice,
-  underlyingAsset,
-}: Record<string, string>) =>
-  `https://www.deribit.com/api/v2/public/get_order_book?instrument_name=${underlyingAsset}-${expiryDate}-${strikePrice}-${optionType}`;
 
 const generateMintAndLiquidateActions = ({
   collateralToDeposit,
@@ -104,7 +81,7 @@ export async function calculateLiquidationTransactionCost({
   return (
     ((
       await gammaControllerProxyContract.estimateGas.operate(
-        generateMintAndLiquidateActions
+        mintAndLiquidationActions
       )
     )
       .mul(gasPriceStore.getLastCalculatedGasPrice())
@@ -112,76 +89,6 @@ export async function calculateLiquidationTransactionCost({
       10 ** 18) * // Ether has 18 decimals, gas is calculated in wei
     (await fetchDeribitETHIndexPrice())
   );
-}
-
-export async function fetchCollateralAssetDecimals(
-  collateralAssetAddress: string
-) {
-  const collateralAssetContract = new ethers.Contract(
-    collateralAssetAddress,
-    erc20ABI,
-    provider
-  );
-
-  return collateralAssetContract.decimals();
-}
-
-export async function fetchDeribitBestAskPrice({
-  expiryDate,
-  optionType,
-  strikePrice,
-  underlyingAsset,
-}: Record<string, string>) {
-  return (
-    await (
-      await fetch(
-        generateDeribitUrl({
-          expiryDate,
-          optionType,
-          strikePrice,
-          underlyingAsset,
-        })
-      )
-    ).json()
-  ).result.best_ask_price;
-}
-
-export async function fetchDeribitETHIndexPrice() {
-  return (
-    await (
-      await fetch(
-        `https://deribit.com/api/v2/public/get_index_price?index_name=eth_usd`
-      )
-    ).json()
-  ).result.index_price;
-}
-
-export async function fetchShortOtokenDetails(shortOtokenAddress: string) {
-  const shortOtokenContract = new ethers.Contract(
-    shortOtokenAddress,
-    oTokenABI,
-    provider
-  );
-
-  return shortOtokenContract.getOtokenDetails();
-}
-
-export async function fetchShortOtokenInstrumentInfo(
-  shortOtokenAddress: string
-) {
-  const shortOtokenContract = new ethers.Contract(
-    shortOtokenAddress,
-    oTokenABI,
-    provider
-  );
-
-  const instrumentInfo = await shortOtokenContract.symbol();
-  const [, expiryDate, strikePriceAndOptionType] =
-    instrumentInfo.match(/([^-]+)/g);
-
-  const [strikePrice, optionType] = strikePriceAndOptionType.match(/\d+|\D+/g);
-
-  return { expiryDate, optionType, strikePrice };
 }
 
 export const marginCalculatorContract = new ethers.Contract(
